Removing compiler warnings.
Removed nearly 100 compiler warnings in the various routing daemons
which now clean compile.

Removed one warning in vty.c, which was obscured by the other
warnings.  SO... this commit corrects the previous one.

+++ ospfd/ospf_packet.c
@@ -606,8 +606,7 @@ ospf_write (struct thread *thread)
    * and reliability - not more data, than our
    * socket can accept
    */
-  maxdatasize = MIN (oi->ifp->mtu, ospf->maxsndbuflen) -
-    sizeof (struct ip);
+  maxdatasize = MIN((int)oi->ifp->mtu, ospf->maxsndbuflen) - sizeof (struct ip);
   
   /* Get one packet from queue. */
   op = ospf_fifo_head (oi->obuf);
